+++ NEWS.md
@@ -4,6 +4,7 @@
 
   * `anomalize()` fails gracefully if 'reference' is not a raster object, and 
     uses the 1st layer (with a warning) if a multi-layer object is supplied.
+  * `denoise()` fails gracefully if neither 'expl.var' nor 'k' are supplied.
 
 #### 🐛 bug fixes