+++ NEWS.md
@@ -1,4 +1,4 @@
-# remote 1.2.3.9009 (2026-07-22)
+# remote 1.2.3.9010 (2026-07-28)
 
 #### ✨ features and improvements